Implement adding tickets to event

+++ Evento.Core/Domain/Event.cs
@@ -31,7 +31,7 @@ namespace Evento.Core.Domain
             UpdateAt = DateTime.UtcNow;
         }
 
-        public void AddTickets(int amount, int price)
+        public void AddTickets(int amount, decimal price)
         {
             var seating = _tickets.Count + 1;
             for(var i=0; i < amount; i++)
